projects_2026/project2_llm_decision_trees/requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-cors==0.8.1
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.13.0
apsw==3.53.0.0
apswutils==0.1.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astunparse==1.6.3
async-lru==2.3.0
attrs==26.1.0
autogluon.common==1.5.0
autogluon.core==1.5.0
autogluon.features==1.5.0
autogluon.tabular==1.5.0
autograd==1.8.0
autograd-gamma==0.5.0
autoprognosis==0.1.22
babel==2.18.0
backoff==2.2.1
beartype==0.22.9
beautifulsoup4==4.14.3
bleach==6.3.0
blis==1.3.3
boto3==1.42.91
botocore==1.42.91
catalogue==2.0.10
catboost==1.2.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorful==0.5.8
colorlog==6.10.1
confection==1.3.3
contractions==0.1.73
cryptography==46.0.7
cymem==2.0.13
dataclasses-json==0.6.7
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
einops==0.8.2
einx==0.4.3
emoji==2.15.0
eval_type_backport==0.3.1
faiss-cpu==1.13.2
fastai==2.8.7
fastcore==1.12.39
fastdownload==0.0.7
fastjsonschema==2.21.2
fastlite==0.2.4
fastopic==1.0.1
fastprogress==1.1.5
fasttransform==0.0.2
flatbuffers==25.12.19
formulaic==1.2.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
future==1.0.0
gast==0.7.0
geomloss==0.2.6
google-api-core==2.30.3
google-auth==2.49.2
google-pasta==0.2.0
googleapis-common-protos==1.74.0
graphviz==0.21
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
hmmlearn==0.3.3
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperimpute==0.1.17
hyperopt==0.2.7
idna==3.11
importlib_metadata==8.7.1
installer==0.7.0
interface-meta==1.3.0
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
json5==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.14.0
krippendorff==0.8.2
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-huggingface==1.2.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langchain-together==0.4.0
langdetect==1.0.9
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
lark==1.3.1
libclang==18.1.1
lifelines==0.30.3
lightgbm==4.6.0
llvmlite==0.47.0
loguru==0.6.0
lxml==6.0.3
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
miracle-imputation==0.1.6
mistune==3.2.0
mkl-service==2.5.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.18.1
nbclient==0.10.4
nbconvert==7.17.1
nbformat==5.10.4
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numba==0.65.0
numpy==2.3.5
nvidia-ml-py==13.595.45
oauthlib==3.3.1
ollama==0.6.1
omegaconf==2.3.0
openai==2.31.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.41.0
opentelemetry-exporter-prometheus==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opt_einsum==3.4.0
optree==0.19.0
optuna==4.8.0
orjson==3.11.8
ormsgpack==1.12.2
pandas==2.3.3
pandocfilters==1.5.1
patsy==1.0.2
pbr==7.0.3
pdf2image==1.17.0
pdfminer==20191125
plotly==6.6.0
plum-dispatch==2.8.0
posthog==7.12.0
preshed==3.0.13
prometheus_client==0.25.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.1.3
py-spy==0.4.1
py4j==0.10.9.9
pyahocorasick==2.3.0
pyarrow==20.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyod==3.2.1
pypdf==6.9.2
pypdfium2==5.7.0
PyQt6==6.10.2
python-discovery==1.2.2
python-dotenv==1.2.2
python-fasthtml==0.13.3
python-iso639==2026.1.31
python-json-logger==4.1.0
python-magic==0.4.27
python-multipart==0.0.26
python-oxmsg==0.0.2
pytz==2026.1.post1
pywinpty==3.0.3
PyYAML==6.0.3
RapidFuzz==3.14.5
ray==2.52.1
redis==7.4.0
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
ruff==0.15.11
s3transfer==0.16.0
safetensors==0.7.0
scipy==1.16.3
Send2Trash==2.1.0
sentence-transformers==5.4.0
setuptools==80.10.2
shap==0.51.0
shellingham==1.5.4
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.49
srsly==2.5.3
starlette==1.0.0
statsmodels==0.14.6
stevedore==5.4.1
sympy==1.14.0
tabpfn==7.1.1
tabpfn-common-utils==0.2.19
tenacity==9.1.4
tensorboardX==2.6.5
tensorflow==2.21.0
termcolor==3.3.0
terminado==0.18.1
textsearch==0.0.24
thinc==8.3.13
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
topmost==1.0.2
torch==2.9.1
torchvision==0.24.1
transformers==4.57.6
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.22.18
unstructured-client==0.43.2
uri-template==1.3.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
virtualenv==21.2.4
wasabi==1.1.3
weasel==1.0.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wheel==0.46.3
widgetsnbextension==4.0.15
win32_setctime==1.2.0
wrapt==2.1.2
xgboost==3.1.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
